chore(mailchimp): moved getInterestsForUser to mailchimp module and started typescripting
